Extract frontMatter collection into NodeProcessor
With the html include and render stages merged, frontMatter collection
must be incorporated into the singular html processing stage, for
frontMatter from <include>s to be collected correctly.

This is also enabled by the new layouts system implementation, which
processes layouts independently from pages. Thus,  we may resolve the
layout of the page even after processing the page's content.
